add Entity generator and price field
