+++ build.gradle
@@ -18,6 +18,10 @@ allprojects {
     }
 
     dependencies {
+        implementation 'uk.co.jemos.podam:podam:7.1.0.RELEASE'
+        implementation 'javax.xml.ws:jaxws-api:2.3.1'
+
+
         testImplementation 'org.springframework.security:spring-security-test:6.0.1'
         implementation 'org.springframework.boot:spring-boot-starter-validation:3.0.4'
         implementation 'org.springframework.boot:spring-boot-starter-web:3.0.1'
